refactor: isolate store schema initialization
