+++ docs/authority-module-inventory.md
@@ -55,7 +55,7 @@ security sensitivity), not size alone.
 | --- | ---: | ---: | --- | --- |
 | [`api.rs`](../crates/coven-cli/src/api.rs) | 36,410 / 12,496 | 1,373,450 | transport dispatch, parsing, validation, transport authority, domain orchestration, persistence access, response mapping | **Highest concentration.** `handle_request_with_runtime_and_authority` remains the central route dispatcher. Session lifecycle, Ward/Threads proposals, and Automations handlers share the module. The proposal path also coordinates a process-wide Ward audit lock, filesystem artifacts, SQLite state, recovery, and response precedence. |
 | [`daemon.rs`](../crates/coven-cli/src/daemon.rs) | 13,210 / 5,833 | 494,030 | local IPC and TCP transport, daemon lifecycle, live session ownership, process supervision, recovery, scheduler startup, telemetry | **High mutable-state risk.** `LiveSessionRuntime` owns the shared live-session registry and launch/shutdown coordination; `serve_forever` acquires lifecycle authority, initializes durable state, starts recovery and schedulers, and accepts concurrent transports. In `handle_http_stream_with_lifecycle`, ordinary API requests not consumed by owner-local lifecycle or mobile local-control handling derive transport authority and enter the central API dispatch. |
-| [`store.rs`](../crates/coven-cli/src/store.rs) | 12,899 / 6,155 | 476,581 | schema and migration, query/command persistence, retention, storage health mapping | **High fan-out.** The production portion exposes 106 crate/public free functions. `initialize_store` delegates schema work to `initialize_store_schema`, but both remain alongside session, handoff, event, Ward, hub, Automations, maintenance, and health queries. The process-global `initialized_store_paths` cache and `STORAGE_HEALTH_SNAPSHOTS` add ownership coupling. |
+| [`store.rs`](../crates/coven-cli/src/store.rs) | 12,899 / 6,155 | 476,581 | schema and migration, query/command persistence, retention, storage health mapping | **High fan-out.** The production portion exposes 106 crate/public free functions. `initialize_store` delegates schema work to its private `schema` module, while session, handoff, event, Ward, hub, Automations, maintenance, and health queries remain in the parent. The process-global `initialized_store_paths` cache and `STORAGE_HEALTH_SNAPSHOTS` add ownership coupling. |
 | [`automations/runner.rs`](../crates/coven-cli/src/automations/runner.rs) | 10,101 / 3,966 | 376,070 | occurrence dispatch, scheduler fencing, runtime adoption, retries, timeout/cancellation, crash recovery, terminal settlement | **High state-machine risk.** `dispatch_occurrence_with_clock` combines direct SQL, scheduler-generation checks, runtime launch, cancellation, and settlement; recovery and stop fencing share the production module. #856 has landed, but remaining #857 correctness work should settle before structural movement. |
 | [`pty_runner.rs`](../crates/coven-cli/src/pty_runner.rs) | 9,954 / 5,551 | 367,600 | harness command construction, PTY/piped I/O, stream decoding, platform process containment, timeout/cancellation, supervisor protocol | **High platform/security risk.** Strict child containment, guardian/supervisor protocol, stream adapters, and detached/attached execution share one module. Windows and Unix ownership guarantees converge at `spawn_strict_child_process_tree`, while cancellation also uses process-global signal coordination in `SUPERVISED_STREAM_CANCELLATION_SIGNAL` and `SUPERVISED_STREAM_CANCELLATION_LOCK`. |
 | [`main.rs`](../crates/coven-cli/src/main.rs) | 9,028 / 5,542 | 330,896 | CLI parsing and command dispatch, setup, user-facing mapping | **Medium.** High fan-in but mostly presentation/entry-point code. Do not prioritize it ahead of authority-bearing modules merely for size. |
